move requireAuth to components
